split monolithic external class into one class per web service
The single block_playerhud\external class held all 15 web service
functions in a 1900-line file. Split it into one class per function
under classes/external/, each following the Moodle 4.x convention of
execute_parameters/execute/execute_returns. The web service names are
unchanged, so db/services.php now points each entry at its dedicated
class with methodname 'execute' and no client (AMD or mobile) is
affected.

The node return structure shared by load_scene and make_choice moved
to a scene_helper class; the four chat-action helpers became private
methods of execute_chat_action.

External web service tests were reorganised under tests/external/ to
mirror the new class layout, sharing fixtures through an
external_base_testcase. Coverage was extended to the previously
untested item-collection, drop-shortcode, story and AI error paths.
