+++ tests/external/generate_ai_content_test.php
@@ -1,0 +1,70 @@
+<?php
+// This file is part of Moodle - https://moodle.org/
+//
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Moodle is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Moodle.  If not, see <https://www.gnu.org/licenses/>.
+
+/**
+ * Tests for the generate_ai_content web service (validation and error paths).
+ *
+ * These tests never hit the network: with no API key configured the generator
+ * fails fast and the web service reports success=false.
+ *
+ * @package    block_playerhud
+ * @category   test
+ * @copyright  2026 Jean Lúcio
+ * @license    https://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+
+namespace block_playerhud\external;
+
+use block_playerhud\tests\external\external_base_testcase;
+use core_external\external_api;
+
+/**
+ * Tests for the generate_ai_content web service.
+ *
+ * @package    block_playerhud
+ * @copyright  2026 Jean Lúcio
+ * @license    https://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ * @covers     \block_playerhud\external\generate_ai_content
+ */
+final class generate_ai_content_test extends external_base_testcase {
+    /**
+     * With no API key the call returns success=false instead of throwing, and
+     * the failure response still validates against the return structure.
+     */
+    public function test_generate_ai_content_without_key_returns_failure(): void {
+        set_config('apikey_gemini', '', 'block_playerhud');
+        set_config('apikey_groq', '', 'block_playerhud');
+        set_config('apikey_openai', '', 'block_playerhud');
+
+        $result = generate_ai_content::execute($this->instanceid, $this->course->id, 'A magic sword');
+
+        $this->assertFalse($result['success']);
+        $cleaned = external_api::clean_returnvalue(generate_ai_content::execute_returns(), $result);
+        $this->assertFalse($cleaned['success']);
+    }
+
+    /**
+     * A student without block/playerhud:manage must be rejected.
+     */
+    public function test_generate_ai_content_requires_manage_capability(): void {
+        $student = $this->getDataGenerator()->create_user();
+        $this->getDataGenerator()->enrol_user($student->id, $this->course->id, 'student');
+        $this->setUser($student);
+
+        $this->expectException(\required_capability_exception::class);
+        generate_ai_content::execute($this->instanceid, $this->course->id, 'A magic sword');
+    }
+}
